reached PHPStan level 3.

+++ api/v1/Kpi.php
@@ -79,7 +79,7 @@ class Kpi extends Base
         int $kpiCountry = -1
     ): int {
         $value = floatval(str_replace(",", ".", $kpiValue));
-        return Database::GetInstance()->query(
+        return (int)Database::GetInstance()->query(
             "
             INSERT INTO kpi (kpi_name, kpi_value, kpi_month, kpi_type, kpi_lastupdate, kpi_unit, kpi_country_id) 
             VALUES (?, ?, ?, ?, ?, ?, ?)
